Prepare for regular docker rebuild and republish
To keep the system and libaries inside the docker container up-to-date
each release of the container needs to be rebuild on a regular basis.
Therefore the pipeline on the tagged commits needs to be rerun regularly.
To avoid also republishing the other artifacts
a variable is introduced to suppress the execution of those jobs.